+++ activerecord/lib/active_record.rb
@@ -464,6 +464,15 @@ module ActiveRecord
     Marshalling.format_version = value
   end
 
+  ##
+  # :singleton-method:
+  # Sets the database connection adapter to be used when the adapter <tt>mysql</tt> is specified.
+  # Accepted values are <tt>:mysql2</tt>, <tt>:trilogy</tt>, and <tt>nil</tt> which does not register
+  # any <tt>mysql</tt> adapter alias.
+  # Defaults to <tt>nil</tt>.
+  singleton_class.attr_accessor :mysql_adapter
+  self.mysql_adapter = nil
+
   def self.eager_load!
     super
     ActiveRecord::Locking.eager_load!